# [Enhancement] adjust translate function syntax structure to prevent parser performance rollback